Commit Inicial, Se agrega la versión 5.7.5.13 con los cambios en la cabecera por marca pais

+++ concrete/controllers/backend/file.php
@@ -2,12 +2,18 @@
 namespace Concrete\Controller\Backend;
 
 use Concrete\Core\File\Importer;
+use Concrete\Core\File\ImportProcessor\ConstrainImageProcessor;
+use Concrete\Core\File\ImportProcessor\SetJPEGQualityProcessor;
+use Concrete\Core\Foundation\Queue\Queue;
+use Concrete\Core\Validation\CSRF\Token;
 use Controller;
 use FileSet;
 use File as ConcreteFile;
 use \Concrete\Core\File\EditResponse as FileEditResponse;
 use Loader;
 use FileImporter;
+use Config;
+use stdClass;
 use Exception;
 use Permissions as ConcretePermissions;
 use FilePermissions;
@@ -34,42 +40,94 @@ class File extends Controller
         $r->outputJSON();
     }
 
+    protected function doRescan($f)
+    {
+        $resize = \Config::get('concrete.file_manager.restrict_uploaded_image_sizes');
+        $processors = array();
+        if ($resize) {
+            $width = (int) \Config::get('concrete.file_manager.restrict_max_width');
+            $height = (int) \Config::get('concrete.file_manager.restrict_max_height');
+            $quality = (int) \Config::get('concrete.file_manager.restrict_resize_quality');
+            $resizeProcessor = new ConstrainImageProcessor($width, $height);
+            $qualityProcessor = new SetJPEGQualityProcessor($quality);
+            $processors[] = $resizeProcessor;
+            $processors[] = $qualityProcessor;
+        }
+
+        if (count($processors)) {
+            $fv = $f->createNewVersion(true);
+            foreach($processors as $processor) {
+                if ($processor->shouldProcess($fv)) {
+                    $processor->process($fv);
+                }
+            }
+        } else {
+            $fv = $f->getApprovedVersion();
+        }
+        $resp = $fv->refreshAttributes();
+        switch ($resp) {
+            case \Concrete\Core\File\Importer::E_FILE_INVALID:
+                $errorMessage = t('File %s could not be found.', $fv->getFilename()) . '<br/>';
+                throw new Exception($errorMessage);
+                break;
+        }
+    }
+
     public function rescan()
     {
         $files = $this->getRequestFiles('canEditFileContents');
         $r = new FileEditResponse();
         $r->setFiles($files);
-        $successMessage = '';
-        $errorMessage = '';
-        $successCount = 0;
+        $error = new \Concrete\Core\Error\Error;
 
-        foreach ($files as $f) {
-            try {
-                $fv = $f->getApprovedVersion();
-                $resp = $fv->refreshAttributes();
-                switch ($resp) {
-                    case \Concrete\Core\File\Importer::E_FILE_INVALID:
-                        $errorMessage .= t('File %s could not be found.', $fv->getFilename()) . '<br/>';
-                        break;
-                    default:
-                        $successCount++;
-                        $successMessage = t2('%s file rescanned successfully.', '%s files rescanned successfully.',
-                            $successCount);
-                        break;
-                }
-            } catch (\Concrete\Flysystem\FileNotFoundException $e) {
-                $errorMessage .= t('File %s could not be found.', $fv->getFilename()) . '<br/>';
-            }
+        try {
+            $this->doRescan($files[0]);
+            $r->setMessage(t('File rescanned successfully.'));
+        } catch (\Concrete\Flysystem\FileNotFoundException $e) {
+            $errorMessage = t('File %s could not be found.', $files[0]->getFilename()) . '<br/>';
+            $error->add($errorMessage);
+        } catch(\Exception $e) {
+            $error->add($e->getMessage());
         }
-        if ($errorMessage && !$successMessage) {
-            $e = new \Concrete\Core\Error\Error;
-            $e->add($errorMessage);
-            $r->setError($e);
-        } else {
-            $r->setMessage($errorMessage . $successMessage);
-        }
+        $r->setError($error);
         $r->outputJSON();
     }
+
+    public function rescanMultiple()
+    {
+        $files = $this->getRequestFiles('canEditFileContents');
+        $q = Queue::get('rescan_files');
+        if ($_POST['process']) {
+            $obj = new stdClass();
+            $messages = $q->receive(5);
+            foreach($messages as $key => $msg) {
+                // delete the page here
+                $file = unserialize($msg->body);
+                $f = \Concrete\Core\File\File::getByID($file['fID']);
+                if (is_object($f)) {
+                    $this->doRescan($f);
+                }
+                $q->deleteMessage($msg);
+            }
+            $obj->totalItems = $q->count();
+            if ($q->count() == 0) {
+                $q->deleteQueue(5);
+            }
+            print json_encode($obj);
+            exit;
+        } else if ($q->count() == 0) {
+            foreach($files as $f) {
+                $q->send(serialize(array(
+                    'fID' => $f->getFileID()
+                )));
+            }
+        }
+
+        $totalItems = $q->count();
+        Loader::element('progress_bar', array('totalItems' => $totalItems, 'totalItemsSummary' => t2("%d file", "%d files", $totalItems)));
+        return;
+    }
+
 
     public function approveVersion()
     {
@@ -87,11 +145,18 @@ class File extends Controller
 
     public function deleteVersion()
     {
+        /** @var Token $token */
+        $token = $this->app->make('token');
+        if (!$token->validate('delete-version'))
+
         $files = $this->getRequestFiles('canEditFileContents');
         $r = new FileEditResponse();
         $r->setFiles($files);
         $fv = $files[0]->getVersion(Loader::helper('security')->sanitizeInt($_REQUEST['fvID']));
         if (is_object($fv) && !$fv->isApproved()) {
+            if (!$token->validate('version/delete/' . $fv->getFileID() . "/" . $fv->getFileVersionId())) {
+                throw new Exception($token->getErrorMessage());
+            }
             $fv->delete();
         } else {
             throw new Exception(t('Invalid file version.'));
@@ -130,8 +195,10 @@ class File extends Controller
             throw new Exception(t("Unable to add files."));
         }
 
-        if (\Loader::helper('number')->getBytes(ini_get('post_max_size')) < $_SERVER['CONTENT_LENGTH']) {
-            throw new Exception(FileImporter::getErrorMessage(Importer::E_FILE_EXCEEDS_POST_MAX_FILE_SIZE));
+        if ($post_max_size = \Loader::helper('number')->getBytes(ini_get('post_max_size'))) {
+            if ($post_max_size < $_SERVER['CONTENT_LENGTH']) {
+                throw new Exception(FileImporter::getErrorMessage(Importer::E_FILE_EXCEEDS_POST_MAX_FILE_SIZE));
+            }
         }
 
         if (!Loader::helper('validation/token')->validate()) {
